Refactor cases completed report
